requirements.txt
chromadb==0.4.24
crewai==0.35.7
faiss-cpu==1.8.0.post1
flair==0.13.1
gensim==4.3.2
fastapi==0.111.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==1.0.4
langchain-groq==0.1.5
langchain-nvidia-ai-endpoints==0.1.2
nltk==3.8.1
ragatouille
pymongo==4.8.0
sentence-transformers==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.9
scipy==1.12
redis==5.0.7
redis[hiredis]
python-docx==1.1.2
docxcompose==1.4.0
PyMuPDF==1.24.7